chore: refactored z3-logic for hskip law proofs into separate dir in utilities dir, i.e. z3-logic, also patched vite dep from npm audit
